+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file. For commit 
    - `CTRL+F2` switch to gallery view.
    - `CTRL+F3` switch to list view.
    - `Double click` on empty space to select all items.
+ - Wildcard search support (limited to sqlite GLOB)
 
  **Notes**:
  - Enhanced indexing scheduler which doesn't wake the disk as often.